+++ HELM/source/package.json
@@ -3,9 +3,6 @@
   "friendlyName": "HELMWebEditor",
   "version": "1.1.4",
   "description": "HELMWebEditor forked from Pistoia Alliance",
-  "dependencies": {
-    "datagrok-api": "1.18.6"
-  },
   "devDependencies": {
     "@datagrok/js-draw-lite": "../../../JSDraw.Lite/source/web",
     "@types/node": "^18.11.18",
@@ -14,6 +11,7 @@
     "css-loader": "^6.7.3",
     "eslint": "latest",
     "eslint-config-google": "latest",
+    "filemanager-webpack-plugin": "^8.0.0",
     "source-map-loader": "^4.0.1",
     "ts-loader": "^9.2.5",
     "typescript": "^4.8.4",
